backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
openpyxl==3.1.5
email-validator==2.2.0
httpx==0.27.2
reportlab==4.2.5
openai==2.29.0
langgraph==1.1.3
pandas==3.0.1
duckdb==1.5.0
